Add TLS servername to transport configuration (#18)
This fixes an issue when the HTTP(S) tunnel is proxying through another
external proxy.  goproxy will try to create a tls.Client when we create
a NewConnectDialToProxy.  tls.Client must be configured with a
tls.Config with either the ServerName, or InsecureSkipVerify, according
to the API documentation: https://golang.org/pkg/crypto/tls/#Client