client.go - fix cancel leak (go vet finding)
We have this finding:

Running go vet on ./rpc
./client.go:213:3: the cancel function is not used on all paths (possible context leak)
./client.go:245:4: this return statement may be reached without using the cancel var defined on line 213

This is due to the cancel being defined, but it's not being called.
Golang has the defer option for clean-up tasks, which should be called right
after defining it. It will then get automatically call when this function ends
(no matter which route).